+++ app/variables.go
@@ -1,20 +1,11 @@
 package app
 
 var (
+	// DB config
+	DB = DBStruct{}
+
 	// ProjectRoot var
 	ProjectRoot string
-
-	// DBHost database host
-	DBHost string
-
-	// DBUsername database user
-	DBUsername string
-
-	// DBPassword database password
-	DBPassword string
-
-	// DBName database name
-	DBName string
 
 	// Verbose logging
 	Verbose bool
@@ -30,9 +21,25 @@ var (
 
 	// OnlyDB runtime variable set with flags
 	OnlyDB bool
-
-	// // DBType database type
-	// DBType string
-	// // IdentityFile SSH key
-	// IdentityFile string
 )
+
+// DBStruct struct
+type DBStruct struct {
+	// Host database host
+	Host string
+
+	// Username database user
+	Username string
+
+	// Password database password
+	Password string
+
+	// Name database name
+	Name string
+
+	// Port database port (as string)
+	Port string
+
+	// Database type (mysql, postgres etc)
+	Type string
+}