+++ cmd/load.go
@@ -51,15 +51,18 @@ var loadCmd = &cobra.Command{
 		app.AddTempFile(assetsFile)
 
 		if utils.IsFile(gzipSQLFile) && !app.OnlyAssets {
-			if err := app.Env(base); err != nil {
-				return err
-			}
-			dropDatabase, _ := cmd.Flags().GetBool("drop-db")
-			if err := utils.CreateDatabase(dropDatabase); err != nil {
+			if err := app.BoostrapEnv(base); err != nil {
 				return err
 			}
 
-			if err := utils.LoadDatabaseFromGz(gzipSQLFile); err != nil {
+			dropDatabase, _ := cmd.Flags().GetBool("drop-db")
+			// use map to determine which database function to use
+			if err := utils.DBCreateWrapper[app.DB.Type](dropDatabase); err != nil {
+				return err
+			}
+
+			// use map to determine which database function to use
+			if err := utils.DBLoadWrapper[app.DB.Type](gzipSQLFile); err != nil {
 				return err
 			}
 		}