+++ cmd/root.go
@@ -43,11 +43,11 @@ func Execute() {
 		app.Cleanup()
 
 		// detect if subcommand is valid
-		help := "\nSee: `ssbak -h` for help"
+		help := "See: `ssbak -h` for help"
 		if len(os.Args) > 0 {
 			for _, t := range rootCmd.Commands() {
 				if t.Name() == os.Args[1] {
-					help = "\nSee: ssbak " + os.Args[1] + " -h` for help"
+					help = "See: ssbak " + os.Args[1] + " -h` for help"
 				}
 			}
 		}